make inbuilt extensions to refit more discoverable
