+++ Refit/Extensions/Properties/DefaultPropertyProvider.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Reflection;
+
+namespace Refit.Extensions.Properties
+{
+    public class DefaultPropertyProvider
+    {
+        public IDictionary<string, object> GetDefaultProperties(MethodInfo methodInfo, Type targetType)
+        {
+            var properties = new Dictionary<string, object> {{HttpRequestMessageOptions.InterfaceType, targetType}};
+
+            return properties;
+        }
+    }
+}
