+++ Refit/RefitSettings.cs
@@ -8,6 +8,8 @@ using System.Reflection;
 using System.Runtime.Serialization;
 using System.Threading;
 using System.Threading.Tasks;
+using Refit.Extensions.Exceptions;
+using Refit.Extensions.Properties;
 
 namespace Refit
 {
@@ -23,7 +25,7 @@ namespace Refit
             UrlParameterFormatter = new DefaultUrlParameterFormatter();
             FormUrlEncodedParameterFormatter = new DefaultFormUrlEncodedParameterFormatter();
             ExceptionFactory = new DefaultApiExceptionFactory(this).CreateAsync;
-            PropertyProviderFactory = new DefaultPropertyProviderFactory().GetDefaultProperties;
+            PropertyProviderFactory = new DefaultPropertyProvider().GetDefaultProperties;
         }
 
         /// <summary>
@@ -41,7 +43,7 @@ namespace Refit
             UrlParameterFormatter = urlParameterFormatter ?? new DefaultUrlParameterFormatter();
             FormUrlEncodedParameterFormatter = formUrlEncodedParameterFormatter ?? new DefaultFormUrlEncodedParameterFormatter();
             ExceptionFactory = new DefaultApiExceptionFactory(this).CreateAsync;
-            PropertyProviderFactory = new DefaultPropertyProviderFactory().GetDefaultProperties;
+            PropertyProviderFactory = new DefaultPropertyProvider().GetDefaultProperties;
         }
 
         /// <summary>
@@ -161,48 +163,6 @@ namespace Refit
                                      ? "{0}"
                                      : $"{{0:{formatString}}}",
                                  enummember?.Value ?? parameterValue);
-        }
-    }
-
-    public class DefaultPropertyProviderFactory
-    {
-        public IDictionary<string, object> GetDefaultProperties(MethodInfo methodInfo, Type targetType)
-        {
-            var properties = new Dictionary<string, object> {{HttpRequestMessageOptions.InterfaceType, targetType}};
-
-            return properties;
-        }
-    }
-
-    public class DefaultApiExceptionFactory
-    {
-        static readonly Task<Exception?> NullTask = Task.FromResult<Exception?>(null);
-
-        readonly RefitSettings refitSettings;
-
-        public DefaultApiExceptionFactory(RefitSettings refitSettings)
-        {
-            this.refitSettings = refitSettings;
-        }
-
-        public Task<Exception?> CreateAsync(HttpResponseMessage responseMessage)
-        {
-            if (!responseMessage.IsSuccessStatusCode)
-            {
-                return CreateExceptionAsync(responseMessage, refitSettings)!;
-            }
-            else
-            {
-                return NullTask;
-            }
-        }
-
-        static async Task<Exception> CreateExceptionAsync(HttpResponseMessage responseMessage, RefitSettings refitSettings)
-        {
-            var requestMessage = responseMessage.RequestMessage!;
-            var method = requestMessage.Method;
-
-            return await ApiException.Create(requestMessage, method, responseMessage, refitSettings).ConfigureAwait(false);
         }
     }
 }
